test
- 9-pMMG에 1,4,7번만 연결하고 통신 확인 완료

+++ TestBoard_G474RE/Core/Src/pMMG.c
@@ -233,17 +233,13 @@ void pMMG_DisableCS(pMMG_Obj_t* pMMG_Obj) {
 }
 
 
-
 void us_Delay(uint32_t us_delay)
 {
-    uint32_t usStart = DWT->CYCCNT / 170;
-    while ( DWT->CYCCNT/170 - usStart < us_delay)
+    float usStart = (float)DWT->CYCCNT / 170;
+    while ( (float)DWT->CYCCNT/170 - usStart < (float)us_delay )
     {
     }
 }
-
-
-
